Added a viewpoint to each point in the point cloud.
In order to generate simulated point clouds for raycasting algorithms, the viewpoint of each point is required. Added additional fields to the buffer array (temporarily storing the point cloud. Each simulated sensor (except the kinect) fills this fields with the ray's point of origin. Be aware that for the sake of simplicity the point of origin corresponds to the sensor position. This is only an approximation, but good enough for most applications.